separated commands into there own folders and fixed up join command and privmsgs with small issues
