+++ src/Commands/Invite.cpp
@@ -1,0 +1,6 @@
+#include "Command.hpp"
+
+void Command::handleInviteCommand(IRCClient* client, const std::vector<std::string>& parameters)
+{
+    std::cout << "Handling INVITE command" << std::endl;
+}
