+++ src/Commands/Mode.cpp
@@ -1,0 +1,6 @@
+#include "Command.hpp"
+
+void Command::handleModeCommand(IRCClient* client, const std::vector<std::string>& parameters)
+{
+    std::cout << "Handling MODE command" << std::endl;
+}
