+++ src/Commands/Part.cpp
@@ -1,0 +1,10 @@
+#include "Command.hpp"
+
+void Command::handlePartCommand(IRCClient* client, const std::vector<std::string>& parameters)
+{
+    if (parameters.empty())
+    {
+        // send error message to client about missing channel name
+        return ;
+    }
+}
