+++ src/Commands/Quit.cpp
@@ -1,0 +1,6 @@
+#include "Command.hpp"
+
+void Command::handleQuitCommand(IRCClient* client, const std::vector<std::string>& parameters)
+{
+    std::cout << "Handling QUIT command" << std::endl;
+}
