+++ src/Commands/Topic.cpp
@@ -1,0 +1,6 @@
+#include "Command.hpp"
+
+void Command::handleTopicCommand(IRCClient* client, const std::vector<std::string>& parameters)
+{
+    std::cout << "Handling TOPIC command" << std::endl;
+}
